requirements.txt
aenum==2.2.4
attrs==20.2.0
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
fbchat==1.9.7
idna==2.10
paho-mqtt==1.5.1
python-dotenv==0.14.0
requests==2.24.0
soupsieve==2.0.1
urllib3==1.25.10
